refactor(app): split follow-up schedule parser
