+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -9,7 +9,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `e2e:playwright` | 48 | 7363 | 100 | A+ | `revert.spec.ts` A+<br>`interaction-audit-targets.ts` A+<br>`interaction-audit-fixtures.spec.ts` A+ |
 | `scripts` | 23 | 4114 | 99 | A+ | `native-desktop-smoke.sh` A+<br>`smoke.sh` A+<br>`packaged_window.py` A+ |
 | `source:QuillCodeAgent` | 51 | 6219 | 100 | A+ | `TrustedRouterOAuth.swift` A+<br>`AgentTypes.swift` A+<br>`MockLLMClient.swift` A+ |
-| `source:QuillCodeApp` | 421 | 47399 | 100 | A+ | `ThreadFollowUpScheduleParser.swift` A+<br>`WorkspaceRemoteProjectToolExecutor.swift` A+<br>`WorkspaceModelComposer.swift` A+ |
+| `source:QuillCodeApp` | 423 | 47434 | 100 | A+ | `WorkspaceRemoteProjectToolExecutor.swift` A+<br>`WorkspaceModelComposer.swift` A+<br>`WorkspaceSurface.swift` A+ |
 | `source:QuillCodeCore` | 29 | 2343 | 99 | A+ | `ModelInfo.swift` A+<br>`AppConfig.swift` A+<br>`AutomationModels.swift` A+ |
 | `source:QuillCodePersistence` | 6 | 354 | 100 | A+ | `ConfigStore.swift` A+<br>`SecretStore.swift` A+<br>`ThreadStore.swift` A+ |
 | `source:QuillCodeSafety` | 7 | 1113 | 99 | A+ | `Safety.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+<br>`StaticSafetyDownloadPolicy.swift` A+ |
@@ -32,7 +32,6 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 
 | Grade | Score | Lines | Module | File | Main Issues |
 | --- | ---: | ---: | --- | --- | --- |
-| A+ | 96 | 494 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ThreadFollowUpScheduleParser.swift` | watch file size (494 lines); 2 lines >120 chars |
 | A+ | 96 | 482 | `test:QuillCodeCoreTests` | `Tests/QuillCodeCoreTests/CoreModelTests.swift` | 29 lines >120 chars |
 | A+ | 96 | 436 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectToolExecutor.swift` | watch file size (436 lines); 1 lines >120 chars |
 | A+ | 96 | 431 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelComposer.swift` | watch file size (431 lines); 2 lines >120 chars |
@@ -72,6 +71,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 97 | 214 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeTopBarSurface.swift` | 45 public declarations |
 | A+ | 97 | 206 | `scripts` | `scripts/native_click_probe_contracts/packaged_window.py` | 13 lines >120 chars; 1 lines >160 chars |
 | A+ | 97 | 205 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/Safety.swift` | 19 public declarations |
+| A+ | 97 | 204 | `source:QuillCodeCore` | `Sources/QuillCodeCore/AutomationModels.swift` | 36 public declarations |
 
 ## Every File
 
@@ -376,7 +376,9 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 37 | `source:QuillCodeApp` | `Sources/QuillCodeApp/SlashWorkspaceCommandParser.swift` | no automated issues |
 | A+ | 100 | 214 | `source:QuillCodeApp` | `Sources/QuillCodeApp/SlashWorktreeCommandParser.swift` | no automated issues |
 | A+ | 100 | 161 | `source:QuillCodeApp` | `Sources/QuillCodeApp/SubagentProgressToolExecutor.swift` | no automated issues |
-| A+ | 96 | 494 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ThreadFollowUpScheduleParser.swift` | watch file size (494 lines); 2 lines >120 chars |
+| A+ | 100 | 372 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ThreadFollowUpScheduleCalendarParser.swift` | no automated issues |
+| A+ | 100 | 85 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ThreadFollowUpScheduleIntervalParser.swift` | no automated issues |
+| A+ | 100 | 72 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ThreadFollowUpScheduleParser.swift` | no automated issues |
 | A+ | 100 | 51 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ToolArtifactDocumentPreviewBuilder.swift` | no automated issues |
 | A+ | 100 | 82 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ToolArtifactImagePreviewBuilder.swift` | no automated issues |
 | A+ | 100 | 114 | `source:QuillCodeApp` | `Sources/QuillCodeApp/ToolArtifactTextPreviewBuilder.swift` | no automated issues |
